% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_hist_overlay.R
\name{spark_hist_overlay}
\alias{spark_hist_overlay}
\title{A histogram with reponse function for a sparklyr table \cr}
\usage{
spark_hist_overlay(sparklyr_table, response_var)
}
\arguments{
\item{sparklyr_table}{is the sparklyr table to pass to the function}

\item{response_var}{is the string response variable you want to overlay the histograms with.}
}
\description{
A histogram for plotting response ove the variables in a table. You will need to reduce the range to around 200 per variable to get effective plots for this method at the moment, as bucketing is not supported (yet).
}
\details{
You must have sparklyr and ggplot2 installed \cr
You must also have the sparkeda jar installed and referenced the same way as spark_hist \cr
You can change the plot output sizes with the chunk settings using knitR like {r fig.height=8, fig.width=20}
\cr
Example selection of a spark table and plot generation\cr
\code{adult_df = tbl(sc, sql("select * from sample_data.adult_dataset"))} \cr
\code{spark_hist_overlay(adult_df, "income"))}
}
