% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_corr.R
\name{spark_plot_corr}
\alias{spark_plot_corr}
\title{A Correlation Matrix Plotting Function for SparklyR \cr}
\usage{
spark_plot_corr(sparklyr_corr_matrix)
}
\arguments{
\item{sparklyr_corr_matrix}{is the sparklyr ml_corr() output you will pass to the function.}
}
\description{
Useful for EDA plotting of ml_on large Spark/Hive tables, it is designed to resemble the ggcorplot2() function in native R. \cr
}
\details{
Example corrplot using sparklyR \cr
\code{spark_cars = copy_to(sc, mtcars)} \cr
\code{s_corr_mat = ml_corr(spark_cars, columns=colnames(spark_cars), method="pearson")} \cr
\code{spark_plot_corr(s_corr_mat)}
}
