% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_log_manLasso.R
\name{spark_plot_log_manLasso}
\alias{spark_plot_log_manLasso}
\title{A SparklyR Logistic Regression Feature Selection Method \cr}
\usage{
spark_plot_log_manLasso(sparklyr_table, predictor, num_folds = 3,
  parallelism = 1)
}
\arguments{
\item{sparklyr_table}{is the spark table you will pass to the function. You can pass using a dplyr spark table (tbl).}

\item{predictor}{is the target column to predict}

\item{num_folds}{(default=3) this param passes the number of cross-validation folds to use for each logistic regression model}

\item{parallelism}{(default=1) this param allows us to deploy default models simultaneously}
}
\description{
The function compares the AUC between logistic regression models for each feature, dropping each iteratively.
}
\details{
Important package requirements: \cr
You must have ggplot2 installed \cr
\cr
Example selection of a spark table and graph\cr
\code{spark_table = tbl(sc, sql("select * from sample_data.iris limit 100"))} \cr
\code{outputs = spark_plot_log_manLasso(spark_table, predictor='Species')}
}
