% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_overlay_pct.R
\name{spark_plot_overlay_pct}
\alias{spark_plot_overlay_pct}
\title{A variable response percentage plot for variables in a sparklyr table \cr}
\usage{
spark_plot_overlay_pct(sparklyr_table, response_var,
  max_numeric_ticks = 40)
}
\arguments{
\item{sparklyr_table}{is the sparklyr table to pass to the function}

\item{response_var}{is the string response variable you want to overlay the histograms with.}

\item{max_numeric_ticks}{40 is the default, using over 40 is fine but you should increase the output width using knitR.}
}
\description{
You will need to reduce the range to around 200 per variable to get effective plots for this method at the moment, as bucketing is not supported (yet).
}
\details{
You must have sparklyr, ggplot2, and purrr installed \cr
You must also have the sparkeda jar installed and referenced the same way as spark_hist \cr
You can change the plot output sizes with the chunk settings using knitR like {r fig.height=8, fig.width=20}
\cr
Example selection of a spark table and plot generation\cr
\code{adult_df = tbl(sc, sql("select * from sample_data.adult_dataset"))} \cr
\code{spark_hist_overlay(adult_df, "income"))}
}
