% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_pairs.R
\name{spark_plot_pairs}
\alias{spark_plot_pairs}
\title{A pairs plot generation tool for sparklyr tables \cr}
\usage{
spark_plot_pairs(sparklyr_table, label_col, sample_size = 80000,
  my_title = "Pairs Plot", progress = FALSE)
}
\arguments{
\item{sparklyr_table}{is the sparklyr table to pass to the function}

\item{label_col}{is the column you want to label in the plot.}

\item{sample_size}{80000 is the default, you can increase this if you wish however it may lead to poor performance.}

\item{my_title}{is the tile of your pairs plot.}

\item{progress}{is FALSE by default, is you change it will produce status bars to let you know the speed.}
}
\description{
This is a lightweight and simple wrapper around GGally for SparklyR tables. It will randomly sample 80,000 records
}
\details{
You must have  GGally, sparklyr, dplyr, and ggplot2 installed \cr
Suggested to store the outputs, and check the chunk settings (as follows) on the stored output to avoid re-running.
You can change the plot output sizes with the chunk settings using knitR like {r fig.height=12, fig.width=12}
\cr
Example selection of a spark table and plot generation\cr
\code{iris_df = copy_to(sc, iris, name="iris_df")} \cr
\code{spark_plot_pairs(sparklyr_table = iris_df, label_col = "Species")}
}
