% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_pca.R
\name{spark_plot_pca}
\alias{spark_plot_pca}
\title{A PCA Plotting Tool for SparklyR ml_pca \cr}
\usage{
spark_plot_pca(ml_pca_model)
}
\arguments{
\item{ml_pca_model}{is the spark pca model outputs table you will pass to the function.}
}
\description{
PCA Plot output with column names. Gives only the top two dimensions
}
\details{
You must have ggrepel installed \cr
\cr
Example selection of a spark table and graph\cr
\code{pca = ml_pca(tbl(sc, sql("select * from db.stock_samples_20m limit 100")))} \cr
\code{spark_plot_pca(pca))}
}
