% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_plot_stddev_group.R
\name{spark_plot_stddev_group}
\alias{spark_plot_stddev_group}
\title{A SparklyR Simple Plotting Method for comparing Data Deviance in Standard Deviations across groups of data \cr}
\usage{
spark_plot_stddev_group(sparklyr_table, group_col)
}
\arguments{
\item{sparklyr_table}{is the spark table you will pass to the function. You can pass using a dplyr spark table (tbl). YOU MAY ONLY INCLUDE 2 COLUMNS}

\item{group_col}{is the target column to predict}
}
\description{
This function creates an easily displayed bar-graph for examining data deviance
}
\details{
Important package requirements: \cr
You must have ggplot2 installed \cr
\cr
Example selection of a spark table and graph\cr
\code{spark_table = tbl(sc, sql("select * from sample_data.iris limit 100"))} \cr
\code{outputs = spark_plot_stddev_group(spark_table, predictor='Species')}
}
